Import Rbots 0.32 sources and build with Maven
Note that this is a best-effort recreation of the original Rbots 0.32 release given the publicly-available artifacts.

A note regarding rbots-ctf-devel: This module corresponds to the rbots-ctf-devel.zip file from the Rbots web site.  However, due to file corruption in all publicly-available copies of this file, some of the content needed to be reconstructed from other available artifacts.  For example, the missing/corrupted *.3do files were replaced with the ones from the SDK, and the missing *.jed files were created by loading the *.jkl files into ZED and saving into *.jed format.  There are obviously some missing *.~je files, but since these are merely backups of the main *.jed files, I did not bother reconstructing those files.

A note regarding rbots-mots-devel: This module corresponds to the rbots-mots-devel.zip file from the Rbots web site.  However, the link on the web site to this file is dead, and there does not appear to be any other publicly-available copy of the file.  So, the contents of this module were reconstructed by taking the contents of the other available artifacts (SDK, episode, etc.).  As with the rbots-ctf-devel module, the *.jed files were created by converting the *.jkl files back into *.jed format via ZED, and no effort was made to restore any *.~je files that may have existed.  Overall, the rbots-mots-devel module in this project should produce a reasonable approximation of the original rbots-mots-devel.zip file.