Overhaul server-side AI, startup, and Rbot Admin cogs
Rbot Admin is a collection of utility cogs that can be used to debug and change configuration settings for a level that is using Rbots.  Since JK/MotS have a very limited number of extra hotkey slots available, Rbot Admin implements its own text-based menu selection mechanism that replaces the individual hotkeys (formerly the "Bot Camera Mode", "Number of bots", and "Bot Difficulty" hotkeys).  The Rbot Admin supports three hotkeys - "Next" and "Prev" to navigate through the available functions, and "Execute" to use the selected action.

Available Rbot Admin functions:

* Add/Remove bots

* Toggle use of guns, sabers, and force powers

* Change difficulty

* Show bots on overlay map

* Toggle slow motion play

* Toggle whether bots will attack players and other bots

* Select next/previous bot - used to select a single bot for various other functions

* Bot Camera - watch the currently-selected bot.  Replaces the "Bot Camera Mode" hotkey.

* Display stats for selected bot

* Kill selected bot

* Highlight thing of interest (node, enemy, powerup) for selected bot (note: unfortunately, the game's targeting reticle doesn't appear when the bot camera is active)

* Teleport selected bot to player's location

* Teleport player to selected bot's location

All of these functions are added as inventory cogs via items.dat.  Apparently, the number of cogs that can be added in items.dat is influenced by the cog counts in static.jkl.  The count has been increased to accommodate all of the new cogs.

The Rbots patch now includes all of the Rbot Admin functionality.  The no-hotkeys patch has not been changed.

The Rbot Server AI cog (rbot-srv.cog) has undergone a significant change.  All of the force power handling has been broken out into a new rbot-srv-force.cog.  There must be one instance of this new cog for each bot.  This migration was necessary to add new functionality to rbot-srv.cog, which had apparantly reached some threshold where adding new code would cause the game to crash when loading the level.  The exact cause of the crash is not clear, but it seemed that the rbot-srv.cog had grown larger than the game was able to handle.

In addition, the bot damage handling code needed to be reorganized, as some of it was directly tied to the use of force powers.  It has been refactored to flow very similarly to the player's own inventory/class cogs in that the rbot-srv.cog first delegates to the rbot-srv-force.cog to perform its handling, possibly reducing the applied damage, and then rbot-srv.cog applies the remaining damage to health/shields accordingly.  Also, some oddities in the force power handling have been resolved and the code now also more closely replicates the behavior of the player damage handling code.

The Rbot Startup cog (rbot-startup.cog) has undergone a massive cleanup and the same cog can now be generically used by all game modes (JK, CTF, and MotS) - that is, each mode no longer has its own "flavor" of the startup cog with minor tweaks.

The startup cog now supports configurable bot respawn masks.  As with the player respawn maske, this is used to control where bots are allowed to respawn.  This is mainly useful in CTF to ensure that bots spawn only in their own team's base, but it can also be useful in normal play if it is necessary to exclude certain player respawn points from being used by bots (e.g. because it would be difficult for the bot to use that spawn point for some reason).

The maxbots parameter has been removed from the startup cog.  The maximum number of bots is now determined at run-time by discovery of rbot-srv.cog instances (via triggers), so it is no longer necessary to configure the count.

The startup cog now supports limited dynamic run-time setup of an Rbots level.  It will automatically make the following adjustments based on the host player's character shortly after starting the level:

* If lightsaber is the selected weapon, switch to sabers-only mode (i.e. disable bot use of guns)

* If fists are the selected weapon, switch to fists-only mode (i.e. disable bot use of guns and sabers)

* If player has no force powers, switch to no-force mode (i.e. disable bot use of force powers)

The Rbot Admin has the ability to change all of these settings at any time after the level starts.  However, this enhancement was designed to provide a limited level of game mode control to the host when the Rbot Admin functions are not available.  For example, many mods will conflict with the Rbots hotkeys patch, so the hsot player would need to play those mods with the no-hotkeys variant of the Rbots patch, which does not include any Rbot Admin functions.

One significant benefit of this enhancement is that it reduces the need for maintaining multiple variants of a level, each with a static permutation of the game settings.  Thus, all unnecessary variants of the Rbots levels have been removed.  This reduces the overhead for maintaining the Rbots levels and also simplifies level selection in-game.  It also adds some flexibility, as some variations (fists only, sabers only) were not previously available for all levels.

In order to make these enhancements, the startup cog will now query rbot-srv.cog for values on-demand (e.g. bot thing references, bot names), rather than listening for triggers and maintaining a cache.  This change was necessary to free up symbols so references to the rbot-srv.cog instances could be maintained for other uses.

All of the aforementioned cogs have received some general formatting and symbol usage cleanup (especially rbot-startup.cog) and now pass parsec.  Many obsolete files have also been removed.

Branding adjustment: episode/level names have been tweaked to remove "final release" branding and to better reflect some of the original JK/MotS names.

Bug fix: For the Valley of the Jedi teams level, bots now use force powers.  This bug was apparently present in the 0.32 release.