Fix rbot-door.cog and rbot-2door.cog to prevent spamming of touched message
If the door is open, a player or bot could spam the touched message by continuously walking into the door, causing it to endlessly try to open.  This was caused by the cog not checking the door's current frame before attempting to move the door.  The cog now only attempts to open the door if it is closed, which repairs the issue.

The rbot-door.cog and rbot-2door.cog files have also been added to the CTF module.