bespin: Bot now open the floor door in the central shaft more reliably
The rbot-door.cog and rbot-2door.cog scripts now treat an actor (bot) walking on the door things as a signal to "activate" the door.  This special treatment is only done for actors to simulate activation of the door; players still need to manually activate the door.

Remove redundant SetTimer call in activate message.  The timer is set in the arrived message, so there is no need to set it in the messages that start door movement.  This was probably a leftover from before the arrived and blocked messages were correctly being sent.

The touched and entered messages now goto the activate message, rather than duplicating the door handling logic.

Also fixed the arrived check in rbot-2door.cog to prevent duplicate timers from being queued that would try to close the doors a second time (after they were already closed).