Use dedicated ghost objects for path nodes
This allows for the possibility of the navigation system to remove all node things after the level starts, which would maximize the number of things that can exist in the level.  This feature is a requirement for adapting larger levels (e.g. The Challenge at Nar Shaddaa, large level), since they will require more nodes and keeping those nodes in the level at runtime would bring the lavel dangerously close to the maximum limit for world things, preventing dynamic creation of things during play.

Although the three levels (arena, duel, freezer) aren't technically in danger of hitting any maximum thing limits, it's still seems more appropriate to have a dedicated node thing instead of hijacking ghosts that need to be in the level for other reasons.