Add dflt_nodesize parameter to nav cog
This parameter can be used to change the default node size that is used for node things that don't set a custom "size" property in the JKL.

Processing to determine node size is now exclusive to the nav cog.  After srv asks nav for the next node to use, nav returns the node size that srv should use.