Refactor CTF path rank/rundir into rbot-pathinfo-ctf.cog
The CTF variant of Rbots now uses the same rbot-path.cog as the JK and MotS variants.  The new rbot-pathinfo-ctf.cog (present only in the CTF variant) acts as an optional extension of rbot-path.cog.  When added to a level, it provides the CTF path ranking and run direction evaluation logic that was previously done in rbot-pathctf.cog, which has been retired.