arena: Tweak item ignoring for reliability
Instead of disabling item seeking on reaching a node, disable it while the bot is in certain sectors.  With the node approach, the bot might get delayed getting up the ramps (e.g. because of combat) and might end up chasing the problematic items anyway.  Using sector entry/exit is more reliable since item seeking will remain disabled until the bot moves out of the problem area.

This tweak specifically addresses a problem where the bot would get distracted by an item while going up the ramps under the CTF bases (before reaching a node that disables item seeking) and then falls down and gets in an endless cycle of trying to get to the node at the top of the first ramp, jumping, getting momentarily distracted by items across the way (the ones we want the bot to ignore), and then continuing to try to reach the same node.  The bot lost timeout is reset every time the bot spots those items, so the bot will truly repeat this cycle forever unless interrupted by an opponent or something like that.  I believe this issue was inadvertently introduced by the previous performance tweaks to the arena level, which moved a few nodes around and may have upset a delicate balance that prevented this issue from occurring so regularly before.