challenge: Performance tuning (node size/placement, item seeking)
Adjust some node placement and sizes.

rbot-srv.cog now performs distance and height checks on items before trying to retrieve them.  This prevents the bot from trying to get items that are too high or too far away.  Also fixed a comment in the item seek to indicate that the 'bot lost' timeout occurs after retrieving the item, rather than before.  Although the intent may have been to give the bot 5 seconds to get an item, the reality is that the "bot lost" timeout is effecitvely suspended until the bot is no longer seeking items.

When bot needs to jump per special node info, there is no longer any velocity added to the bot.  It didn't seem to be having any real effect, which might be caused in part by the change for the bot to move to 0.25 JKU above the node position to reduce surface drag.