Fix defect causing sector special info to be ignored if bot spawns in the special sector
There was a defect where the special info (e.g. ignore items) for a sector would be ignored by a bot if the bot were to respawn in the sector.  The bot would not honor the special info until exiting and re-entering the sector.  The cause of this defect was order of initialization code execution.  Namely, the rbot-srv2.cog would reset the botmode symbol after the rbot-special-sector.cog had sent the trigger to alter the botmode.

Creation of the bot thing (i.e. with CreateThing* verbs) generates an entered message for the sectors that the bot is created in.  The CreateThing* verbs used to create the bot block until this message returns, meaning that rbot-special-sector.cog must run while rbot-srv.cog is waiting for CreateThing* to return, and therefore before it even has a reference to the created bot thing or has a chance to capture it.  To work around this ordering problem, rbot-special-sector.cog uses a timer to delay the special info trigger if the actor that generated the entered message does not (yet) have a capture cog.  By the time the thread returns (through CreateThing* and rbot-srv.cog) and executes the timer immediately afterward, the bot will have its capture cog set and be fully initialized.

The same defect existed for rbot-special-surface.cog when bots were immediately attached to floor surfaces upon creation.  The same solution (i.e. delaying the special info trigger until srv/srv2 initialization is complete) also repaired the issue for surface special info.