Add special mode to disable "combat roaming" (i.e. moving to engage/follow target)
When the 0x8000000 flag (disable combat roaming) is enabled in the bot mode, then the bot will not stray from its current path to engage/follow an enemy.  The bot will still attack any enemies encountered along the path, but the bot will move directly to the next path node and will not strafe around the enemy or try to follow/head directly for the enemy.

This new bot mode flag is separate from 0x4000000 (disable item seeking) to provide level developers with maximum flexibility for controlling how and when bots are allowed to stray from the path.  For example, a developer may want to prohibit a bot from seeking items in a particular area, while engaging an enemy in prolonged, off-path combat in that same area might be fine.

This new bot mode flag can be set/cleared with the new 0x800 special mode flag.

The rbot-srv2.cog has once again reached the reference limit, so the user7 message (used with SendMessageEx based API) has been simplified to omit superfluous 'else' keyword and error message usage.