Bots emit extra light when lightsaber is active
When a bot selects the lightsaber, it is given extra light to emulate the glow that would normally be emitted from the engine-rendered lightsaber blade.  The light=0.65 value used is an approximation of how much light the engine-rendered blade would have based on visual comparison between bots and players (i.e. this value was not obtained through any exact scientific means).

Note: Normally the player has light=0.2 when the lightsaber is not mounted so there is a bit of light on walls, etc. as the player gets very close to them.  It is certainly possible to emulate this with bots, but since the effect is so subtle, it makes more sense to forego this effect and simply reset the bot's light to zero when the saber is dismounted.  This prevents the bot thing from needlessly consuming a valuable dynamic light resource (the engine has a relatively low limit on these), which effectively yields priority to other dynamic light sources when the bot is not using a saber.  One could argue that a bot consuming a dynamic light resource is no worse than a player consuming that same resource (bots are just substitutes for players), but again, the effect is so subtle, so it makes more sense to allow other more noticeable effects (e.g. weapon projectiles) to use the dynamic lights instead when possible.