Improve test.bat scripts to be more user-friendly
The scripts will assume default install path if the JK_HOME/JKM_HOME variable is not set.

The scripts will also display a nice error message if the game executable is not found. To aid in troubleshooting, the message will include the full path to the executable it is trying to execute.