Remove unnecessary conditional setting of armed mode
The code was checking the current major mode to see if the bot was on land (0-2) or in water (3-5) and then setting the armed mode to the value of the desired major mode. However, this is redundant, since SetArmedMode() will automatically convert the given armed mode value to the appropriate major mode based on whether the bot is in water or not.

The removal of these unnecessary conditions simplifies the code and also reduced the number of code branches (if-else blocks) that count toward the pointer limit.