Fix pit fall detection
When a bot is in a sector that is flagged as a pit, the bot will now only be killed if it is falling fast enough.

For players who are falling in a pit sector, the game engine appears to automatically fade to black and kill them once the player's falling speed exceeds 3.0. The bot pit check likewise inspects the bot's falling speed and will kill the bot if it exceeds 3.0.

This change fixes certain cases where a bot is legitimately allowed to exist within a sector flagged as a pit, such as the elevator shaft in the throne room level.