throne: Re-convert throne room from vanilla JKL
This re-conversion corrects some discrepancies between the original version and Rbots version of the level, including location of the revive pack and the elevation of the catwalks.

Due to an apparent bug in ZED, the extralight setting for several surfaces was not preserved in the conversion (which explains why this discrepancy also existed in the original Rbots version of the level). Affected surfaces include the green view screens on either side of the Emperor's throne and the ceiling between the elevator shaft and the stairs. The extralight for these surfaces has been manually corrected to match the original level.

Placement and special info for some nodes have been adjusted for the difference in catwalk elevation. All cogs have been updated with the new thing numbers, since the previous version of the level had apparently removed some things that were present in the original level (I believe these were the support columns for the catwalks).

The elevator behavior has been adjusted to behave more like the original level. Namely, the elevator will now move, wait, and return automatically, rather than waiting at the upper floor indefinitely.