Bots now handle player carbonite gun attacks properly
The player can now use the carbonite gun to freeze and kill bots. The bots no longer attempt to move/jump while frozen and will properly respawn after being frozen/killed by the carbonite gun.

Previously, the convention had been for the srv cog to capture the bot thing to receive notification of damaged messages. However, this is problematic with the carbonite gun, which also captures things that it freezes. A thing may only have one capture cog, so when the carbonite gun was used, the linkage to the srv cog was severed, and as a result, the srv cog could no longer detect when the bot was damaged or killed. This has been addressed by setting the srv cog as the bot's class cog, rather than as its capture cog. This permits the carbonite gun to capture/release the bot things freely, so both the weapon and srv cogs will receive the damaged messages. This also has the benefit of allowing the srv cog to implement a skill message for the bot to handle force powers, etc., as the skill message is only sent to a thing's class cog and not to its capture cog (one of the few differences between a class cog and a capture cog). Also, it is unlikely that game code or mods would change the bot's class cog. It's also fitting for the srv cog to be the bot's class cog, since it serves a very similar purpose as the kyle.cog does for a player.

The bot isn't really aware of whether it is frozen or not, so the srv cogs check for the 0x2000 AI mode flag (AI disabled) to decide whether to try and move, attack, jump, etc.

All cogs that had previously queried the bot's capture cog to retrieve information about the bot have been altered to query the bot's class cog instead.