Add collision type check when picking up powerups
This check mitigates an issue where bots can immediately recover force-pulled weapons before the powerup collision is reset. Sometimes, the powerup would still be within reach of the bot with collide=0 when the item pickup code is executed by the srv2 pulse. Force pull sets collide=1 shortly after the powerup is thrown (to prevent the victim from touching the powerup and recovering it). Bots now also honor the collide type when checking for powerup pickup.