challenge: Add recovery path to mitigate bots waiting indefinitely after missing elevator arrival
For the elevator leading up to the gold base, bots may get stuck waiting for the elevator if its arrival event was missed or never sent. For example, if a bot is stuck in the elevator shaft, then the elevator would be blocked and sent back to the upper frame without generating an arrival event. A short recovery path has been added so when a bot times out awaiting arrival of the elevator, he will walk out of the room and then back in to trigger the elevator to move again, rather than simply waiting at the same node continuously. This recovery path is unidirectional and will only be used when a bot is lost (e.g. after the arrival timeout). The other nodes in the room have been flagged with the 0x1 nav special flag to force the bots to ignore the normal path when lost and to always use the recovery path in this area.