bes: Add recovery paths to get bots out from under the elevators in the lower area
When the bot gets stuck under one of the elevators in the lower level and becomes lost due to navigation timeout, the bot will use a special recovery path to raise the elevator (if needed) and jump out of the shaft. The path is unidirectional and can only be reached by the bot getting lost in the bottom of the elevator shaft. This mitigates the bots getting stuck under the elevator and then committing suicide due to excessive lost count threshold.

Also optimize special info for upper area to flag the entire sector with 0x2/slow movement flag, rather than flagging the individual nodes.