Adjust bot jump height for consistency
All cases that cause the bot to jump now use consistent height jump scaling (including special info jump, dodge jump, and obstruction jump).

Jump scaling is now also more consistent with the ability of an actual player. Scale of 2.0 is too low, and scale of 3.0 is too high (bot can jump over the large red crates in valley level). Scale of 2.5 seems to be approximately equivalent to an actual player. Scale of 3.0 also broke precision jumps in various levels, but the same jumps still perform correctly most of the time with scale of 2.5.

oasis: Removed an unnecessary 0x4/jump special info from the node at the top of the ramp by the lightning trap. I believe this is meant to get the bot over to the left area by the sequencer powerups, but it really should be implemented better since it also causes the bot to jump when simply going down the ramp. I'm not certain how consistently it worked anyway. One tradeoff, however, is that now the bot relies more on the obstruction jump to reach that laft area.

bgj: Adjusted positions and sizes of several path nodes inside the crashed cargo ship area to improve navigation reliability. Combined two shorter paths into one long looping path around the center area of the level. Removed an unnecessary 0x200/extra speed special info from a path node leading up to the jump across the gap in the eastern area of the level.