Adjust 0x200/extra speed special info for more stable results
The extra speed was previously based on scaling the bot's current velocity (not normalized). If the bot is not simply running in a straight line at normal speed, this can lead to unstable or unexpected results. For example, in bgj, the 0x200/extra speed and 0x4/jump special info flags were used together at the node where the bot jumps across the gap in the eastern area of the level. The bot's velocity included a significant upward component because of the jump, so the extra speed scaled this result, wildly hurling the bot far up into the air, sometimes killing the bot by fall damage or collision with the cliff side.

The extra speed is now calculated such that it is always directed toward the next path node, regardless of which direction the bot is moving or facing. The extra speed factor is also smaller now, since the extra speed should only be used to help the bot navigate tough situations, not to deliberately out-perform human players. These changes should help improve the extra speed function to yield more reliable and stable results, especially when combined with other functions such as jumping.

bes: Add 0x200/extra speed to jump nodes on elevator recovery paths. The consistency of the bot's jump was impaired by a previous commit that reduced the magnitude of the bot jump. Although the jump is now more consistent with actual players, the bot more frequently failed to jump out of the elevator shaft after the change. This commit leverages the revised 0x200/extra speed special info to increase the bot's jump height so he can more reliably exit the elevator shaft.