admin: Add ability to save a location for teleport
By default, teleporting a bot/player will use the player's/bot's current location, respectively. The player may optionally toggle the teleport mode to save the player's current position as the teleport location, and then bots/players will be teleported to that location instead. This will aid in setting up test scenarios where the bot needs to be in a particular location but the player needs to be in a different location (e.g. to observe the bot's behavior from other angles).

Teleporting will now also "stop" the bot/player to remove any current movement velocity. This mitigates annoyances with teleporting the bot to the player where, depending on the bot's current movement direction, the bot's inertial movement after the teleport will move him behind the player and get him stuck, forcing the player to move out of the way. Stopping the bot during the teleport enables the bot to run away from the player in any direction without the player getting in the way.