bes: Bots ignore powerups under elevators that could easily crush them
There are powerups in the elevator shafts leading up from the main floor to the upper floor. The bots now ignore these powerups because running into the elevator shaft to get them will call the elevator to the main floor, which will usually crush the bot and kill him before he becomes lost/obstructed and is able to jump out of the elevator shaft.