Move blind bots to a random position that periodically changes
Blind bot movement was previously incomplete. There was a blindvec symbol and a timer block to update it periodically, but the timer was never being set, so I think the bots would just always move toward '0 0 0' when blinded.

The blindvec update timer is now started on bot respawn and stopped on bot death. It periodically calculates a new random blindvec for the bot to move to when it is blinded.