Fix special info handling so bot won't jump in midair
If the bot is in the air when a 0x4 or 0x8 "jump" special info trigger is received, then the srv cog will set a flag to jump at the next opportunity (e.g. after landing on the ground) instead of "jumping" while still in midair.

The special info jump still hand;es the jump request immediately whenever possible to minimize breakage to existing level design. That is, if the bot were to needlessly delay jumping until the next srv cog pulse, the delay may break the bot's ability to reliably make precision jumps. The special info and node placement would have been set up with specific timing in mind, and delaying the jump, even for 1/10 of a second, is enough to break the timing in some cases (e.g. valley and bgj).