spaceport: Improve navigation through area with tower of crates
A few nodes have been added to the tower of crates to implement more efficient jumping to climb the tower. Each node will make the bot jump, so there is no longer reliance on the obstruction jump to climb the tower. Node placement has been adjusted to minimize the bots falling off of the tower of crates. Also, the node at the bottom of the tower is flagged with 0x2/slow special info, which helps mitigate the bots falling off the tower by over-jumping or moving too quickly.

A node has also been inserted into the path behind the door so the bot will try to move through the center of the doorway while the door is opening. This mitigates the bot getting "stuck" in front of the door by trying to walk through the side of the doorway, since the door sticks out somewhat when it is open and does not recede completely behind the wall.

The path now also goes through the double doors on the upper level and down into the main area. This enables the bot to jump down after reaching the upper room instead of simply reversing direction and going back down through the room with the tower of crates. The path has also been flagged as unidirectional, since the bot can no longer traverse it backwards (i.e. the bot cannot jump up from the main area to the double doors; he must climb the crates).