Improve force persuasion effects when bot changes weapons (JK)
In the JK versions, when a bot is using force persuasion and selects a new weapon, the skin cog changes the bot thing's model for the new weapon. However, this resets the bot thing's current geo mode, which causes the bot to become visible momentarily. This "blink" is then corrected by the next "pulse" that updates the geo mode for the bot based on the player's force seeing activation. This has been improved to avoid the "blink" altogether by preserving the bot's current geo mode when changing the model.

Note: This fix is not necessary in the MotS version because bots simply change their weapon mesh, not the entire model. Thus, the bot's current geo mode is never inadvertently reset.