Fix MotS backpack bin numbers
The MotS version of the backpack cog was using the JK weapon bin numbers. Apparently this does work partially, probably due to bin number magic in the MotS exe, but with some issues. For example, the bowcaster and manual sequencer charge weapon bins don't get updated correctly this way. This commit adjusts the MotS backpack cog to use the bin numbers from items.dat and also to correctly enable manual mode for sequencer charges if any are in the backpack.