Harmonize _rbotactor template with walkplayer template
The _rbotactor template has been adjusted to more closely match the walkplayer template. Instead of extending _humanactor and reverting its changes to _actor, _rbotactor now extends _actor directly and adds the same parameters that walkplayer does. Except for a few minor differences, which are needed for technical reasons, _rbotactor is now an actor-based clone of walkplayer. There is one difference between the templates that is not intuitive: the omission of the staticdrag property. If this parameter is added to the template, the bot is unable to start moving on its own from a complete stop until something gives the bot some initial motion (e.g. obstruction jump).

Reimplement checkspeed to more accurately emulate player movement speed using AiSetMoveSpeed instead of AddThingVel. Previously, using AiSetMoveSpeed to set the bot movement speed to 2.0 didn't work properly because the _rbotactor template did not specify appropriate values for the maxvel and maxthrust properties. These differences have been resolved thanks to the harmonization of the _rbotactor template with the walkplayer template, so AiSetMoveSpeed now has the desired effect.

Fix edge case with force speed adjustment of bot movement speed. The rank variable was not being recalculated in this case and therefore could contain the wrong value with unlucky timing if the player uses a force power against the bot.