Cleanup: Reorder placed cogs to hopefully minimize future diffs
Cogs that are not added/removed often (e.g. level cogs, startup, srv) are placed higher in the cog list, while cogs that tend to be added/removed more frequently (e.g. path, special) are placed later in the cog list. In a JKL, inserting or removing a cog will also affect the reference numbers for all cogs below it in the list. This ordering will hopefully make future JKL file diffs smaller, since the cogs that are most likely to be added/removed are now at the end of the list, so all the cogs that don't tend to change will no longer need to appear in the diff just because of a reference number change.

This commit contains no (intentional) changes in functionality.