Fix broken model for rbot_respawn when running the episode GOB
The rbot_respawn template used respawn.3do for its model. This model essentially contains no actual geometry and is meant to be invisible. This works when running from source (e.g. from a *-devel directory), but the model apparently wasn't working when running from the episode GOB. All rbot_respawn things would appear as dflt.3do instead as if the model would not load. The issue is resolved by extracting the respawn.3do from the GOB, so the issue has something to do with finding/reading the model file from the GOB. However, since the model isn't needed anyway, it has been removed entirely in order to resolve the issue. Also, rbot_respawn is now based on _ghostdecor, which more accurately reflects the intended no-collide nature of the respawn point.