ctf: Re-implement CTF extensions as a general customizer mechanism
The rbot-srv-<extension>.cog mechanism has been refactored into a more general customizer mechanism. While the extension mechanism was designed for one rbot-srv-<extension>.cog instance per bot, the new customizer mechanism allows one customizer cog (Type ID 3000) to be shared by all of the rbot-srv.cog instances. The customizer is invoked directly using functions implemented in the user7 message. The bot thing is always one of the parameters, so the customizer will know which bot is involved in each request. The customizer is consulted after all other rbot-srv-*.cog scripts so it has the final say on any queries it supports and may, therefore, override any decisions made by the standard srv scripts.

Rbots support in ctf_main.cog has been re-implemented, starting from the vanilla LEC version of the script. The revised implementation takes care to minimize changes from the original file to yield a cleaner and less noisy diff (ignoring whitespace). The ctf_main.cog also acts as the customizer in CTF matches, replacing the use of rbot-srv-ctf.cog.

The rbot-startup.cog now fully honors the 0x4 MP mode flag. If the flag is enabled for custom score management (e.g. CTF), the startup cog will still show "was killed by" messages, but the default score changes (+1 for kills, -1 for suicides) and score print/display behavior are suppressed.

Bot scores in CTF are now updated in the same manner as player scores. That is, actions such as stealing the flag or killing a flag carrier now award points to the bot or the team just as if a player performed the action.