Check for node LOS after seeking items
If the bot loses LOS with the node after seeking items, the bot will send a 'bot lost' trigger to find the nearest node to get back on track. This will help improve bot navigation in maps where the bots are more likely to seek items and stray far from the path. For example, in duel, the bots have a tendency to seek item after item and go around corners, making it nearly impossible to return to the current node.