kfy: Initial implementation of KFY
Ruins of the Sith is the only supported map so far. It was produced by combining the contents of the MotS k4.jkl file by LEC and the Rbots temple-1.jed file. Several ghost things were added to the original temple-1.jed in a layer named 'padding' to keep the node thing numbers consistent between the two files. This simplifies copy/pasting the nodes and cogs between the files. These padding nodes have been given the 0x18000 thing flags to remove them from the map and free up those thing slots when the game loads.

Some adjustments have been made to the existing paths to improve navigation and to better support KFY navigation. Several path nodes have also been added for more complete coverage of the map. The rbot-path.cog for KFY is an enhanced version of the original that allows specification of path priority and run direction for efficiently finding the Ysalamiri home. It is similar to the rbot-pathinfo-ctf.cog used in CTF except that these extra attributes are entered directly into the KFY path cog rather than a separate pathinfo cog.

Known issue: Ysalamiri Force bubbles are not yet properly supported by the bots. The bots are able to use Force powers even when they or their target are within a Force bubble.

New feature: Dynamic bot team assignment

Team skins are now specified in rbot-startup.cog. When team play is enabled, and if the team skins have been configured, the startup cog will configure the srv cogs with the team and skin to use. The teams will be populated as evenly as possible (e.g. bots 0,2,4,... to red, bots 1,3,5,... to gold). Up to four teams may be configured. If no team skins are configured, then the bots will use their ordinary skins. KFY allows each bot/player to use their own skin; team skins are not used.

Additionally, the bot team assignment may be dynamically changed during gameplay. KFY leverages this to control whether the players not carrying the Ysalamiri may damage one another or not. That is, in Friendly Mode, KFY puts those players on the same team so the friendly fire check will prevent damage except when battling the player carrying the Ysalamiri.

Misc. bugs fixed:

Fixed a bug in rbot-srv.cog that prevented bots from attacking bots and players with team=0 (i.e. not on any team) in team play. Anyone may damage a player whose team is 0; bots should behave the same way. KFY depends upon this behavior for non-friendly mode.

There is a bug in the original LEC implementation of KFY where the game will play the wrong sound when sending the Ysalamiri home. This bug has been fixed.