ctf: Fold rbot-pathinfo-ctf.cog into rbot-path.cog
The CTF rank and run direction information is now passed directly into rbot-path.cog. This is the same pattern that is used by the KFY variant of rbot-path.cog.. The rbot-pathinfo-ctf.cog is no longer needed and has been retired.

All variants of rbot-path.cog have been harmonized for consistency.

The path info query/response triggers have been retired from all variants of the rbot-path.cog. If an extensible path info mechanism is again needed in the future, it should be reimplemented with messages rather than triggers for improved performance.