temple: Fix item special to ignore force surge
The rbot-item-special.cog entry was copied from the KFY variant of temple-1.jed to the MotS variant, but I forgot to verify the thing numbers. The thing numbers are different between the two files, so the MotS variant was ignoring a power cell instead of the force surge. This mistake has been corrected.